requirements.txt
torch>=2.2
torchaudio
modelscope>=1.11.0      # 含 3D-Speaker 预训练权重 :contentReference[oaicite:0]{index=0}
onnxruntime-gpu         # ONNX 加速 :contentReference[oaicite:1]{index=1}
faiss-gpu               # GPU+CPU 混合索引 :contentReference[oaicite:2]{index=2}
soundfile
tqdm
sounddevice
soundfile
pydub              # 用于多格式转 wav，需要 ffmpeg